requirements.txt
h5py
paddlepaddle
opencv_python
matplotlib
numpy
pillow
tqdm
tensorflow